pyproject.toml
[project]
name = "antkeeping-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "crispy-bootstrap5>=2024.10",
    "django~=5.2.0",
    "django-autocomplete-light>=3.12.1",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.3",
    "django-environ>=0.12.0",
    "django-extra-fields>=3.0.2",
    "django-filter>=25.2",
    "django-redis>=5.4.0",
    "django-taggit>=6.1.0",
    "django-tinymce>=4.1.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sorl-thumbnail>=12.11.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "django-debug-toolbar>=5.0.1",
    "djlint>=1.36.4",
    "ruff>=0.14.13",
]
